+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@prisma/client": "^6.12.0",
     "axios": "^1.10.0",
+    "bullmq": "^5.2.0",
     "cors": "^2.8.5",
     "dotenv": "^17.1.0",
     "express": "^5.1.0",
